requirements.txt
# pip install -r requirements.txt
# The code above is what you put in the terminal to download the associated packages necessary

# Core Requirements
torch>=1.8.0
transformers>=4.18.0
numpy>=1.19.0
Pillow>=8.0.0

# CLIP and Image Processing
git+https://github.com/openai/CLIP.git
pycocotools>=2.0.0

# Utilities
tqdm>=4.45.0
requests>=2.25.1
pandas>=1.2.0

# SSL Handling
pyopenssl>=20.0.0

# Additional Packages from Your Imports
enum34; python_version < '3.4'  # For Enum support in older Python
typing-extensions>=3.7.4  # For modern typing support